+++ test/web/core.js
@@ -134,6 +134,23 @@ success
 }
 */
 
+// local request timeout
+
+done = false;
+startTime = Date.now();
+var res = local.dispatch({
+  method: 'get',
+  url: 'httpl://test.com/timeout',
+  timeout: 1000
+});
+res.then(printError, printSuccess).always(finishTest);
+wait(function () { return done; });
+
+/* =>
+success
+{body: "", headers: {}, reason: null, status: 0}
+*/
+
 // == SECTION core - worker local requests
 
 // successful local requests